requirements.txt
networkx
dwave-ocean-sdk
numpy
scipy
opt-einsum
matplotlib
quimb
autoray